+++ utf8.h
@@ -21,6 +21,6 @@ Rune utf8_decode(Slice s, Slice* rest);
 // Encode a Unicode Rune into a Buffer (therefore, containing Bytes).
 // Return the number of bytes the Unicode Rune was encoded into.
 // If the Rune was not valid Unicode, return 0.
-Byte utf8_encode(Rune r, Buffer* b);
+unsigned int utf8_encode(Rune r, Buffer* b);
 
 #endif
